Move CLLocationCoordinate2D extension
